Time: 31 ms (36.98%) | Memory: 61.8 MB (5.71%) - LeetSync

+++ 518-coin-change-ii/coin-change-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     static const int MOD=1e9+7;
-    unsigned long long change(int amount, vector<int>& coins) {
+    int change(int amount, vector<int>& coins) {
         int n=coins.size();
         vector<vector<unsigned long long>> dp(n+1,vector<unsigned long long>(amount+1,0));
         for(int i=0;i<n+1;i++)  dp[i][0]=1;
